+++ src/main/java/com/drmangotea/tfmg/registry/TFMGItems.java
@@ -29,6 +29,7 @@ import com.drmangotea.tfmg.content.items.weapons.lithium_blade.LithiumBladeItem;
 import com.drmangotea.tfmg.content.items.weapons.quad_potato_cannon.QuadPotatoCannonItem;
 import com.drmangotea.tfmg.content.machinery.misc.winding_machine.SpoolItem;
 import com.drmangotea.tfmg.content.machinery.oil_processing.OilHammerItem;
+import com.drmangotea.tfmg.content.machinery.vat.industrial_mixer.mode.MixerMode;
 import com.drmangotea.tfmg.registry.TFMGTags.Items;
 import com.simibubi.create.AllBlocks;
 import com.simibubi.create.AllTags;
@@ -149,8 +150,8 @@ public class TFMGItems {
                     .register(),
             CINDERFLOURBLOCK = REGISTRATE.item("cinderflourblock", Item::new).register(),
             NAPALM_POTATO = REGISTRATE.item("napalm_potato", Item::new).register(),
-            MIXER_BLADE = REGISTRATE.item("mixer_blade", Item::new).register(),
-            CENTRIFUGE = REGISTRATE.item("centrifuge", Item::new).register(),
+            MIXER_BLADE = REGISTRATE.item("mixer_blade", Item::new).properties(p -> p.component(TFMGDataComponents.MIXER_MODE, new MixerMode.Stored(TFMGMixerModes.mixing))).register(),
+            CENTRIFUGE = REGISTRATE.item("centrifuge", Item::new).properties(p -> p.component(TFMGDataComponents.MIXER_MODE, new MixerMode.Stored(TFMGMixerModes.centrifuge))).register(),
             CRANKSHAFT = REGISTRATE.item("crankshaft", Item::new)
                     .model((c, p) -> p.withExistingParent(c.getName(), TFMG.asResource("item/crankshaft_model")))
                     .register(),